Build script outputs hashes.

+++ build.ps1
@@ -37,7 +37,13 @@ if ($Release)
     Copy-Item LICENSE $outDir
     
     Remove-Item -LiteralPath $ReleasePath -ErrorAction SilentlyContinue
+
+    Write-Output "
+## Checksums
+| Architecture | Checksum |
+|---|---|" > checksums.md
 }
+
 # Build for each architecture
 Foreach ($arch in $Architectures)
 {
@@ -47,8 +53,10 @@ Foreach ($arch in $Architectures)
     {        
         go build -ldflags -H=windowsgui -trimpath -o $outDir\winssh-pageant.exe
         if ($LastExitCode -ne 0) { $returnValue = $LastExitCode }
-        # Remove-Item -LiteralPath $ReleasePath -ErrorAction SilentlyContinue
         Compress-Archive -Path $outDir\* -DestinationPath $releaseDir\$ReleasePath-${ver}_$arch.zip -Force
+
+        $hash = (Get-FileHash $outDir\winssh-pageant.exe).Hash 
+        Write-Output "| $arch | $hash |" >> checksums.md
         
         Remove-Item -LiteralPath $outDir\winssh-pageant.exe
     } else {
@@ -56,12 +64,15 @@ Foreach ($arch in $Architectures)
     }
 }
 
-
 # Restore env vars
 $env:GOOS = $oldGOOS
 $env:GOARCH = $oldGOARCH
 
 # Cleanup
-Remove-Item -LiteralPath $BuildPath -Force -Recurse -ErrorAction SilentlyContinue
+if ($Release)
+{
+    Write-Output "" >> checksums.md
+    Remove-Item -LiteralPath $BuildPath -Force -Recurse -ErrorAction SilentlyContinue
+}
 
 exit $returnValue
